Moved load cell out a little to help screw alignment on the connector.
Corrected NFC PCB holes. Added option for holes to be either through the spool holder for nuts or not for press fit inserts.

Split out the original reel holder as openScad was not working correctly with the changes being made, trimmed spool holder seperatly then adding parts to it.

Added chimney option for spool holder connector/mount to give it a bit of a train look.